Fix convert bug

+++ algoTrader.py
@@ -355,8 +355,8 @@ class RealTimeBot:
                     print("="*40)
                     
                     # Calculate ticks (ORIGINAL CALCULATION)
-                    stop_loss_ticks = (self.stop_loss - self.entry_price) / tick_size
-                    take_profit_ticks = (self.entry_price - self.profit_target) / tick_size
+                    stop_loss_ticks = int((self.stop_loss - self.entry_price) / tick_size)
+                    take_profit_ticks = int((self.entry_price - self.profit_target) / tick_size)
                     
                     # Place order with ORIGINAL parameters
                     await self._place_order(1, stop_ticks=stop_loss_ticks, take_profit_ticks=take_profit_ticks)
